NTLM MIC calculation draft
We run into issue where some server type (EMC Unity) fail when MIC was presented and not requested.
Other issue with other type (NetApp 7mode) if MIC  was presented and  not calcullated correctly.

MIC (16 bytes) provided and offset is modified if in AvPairType is key MsvAvFlags with value & 0x00000002 is true
otherwise MIC is skipped and offset is not modified.
MIC calculation is a draft and need further testing.
Documentation on address https://msdn.microsoft.com/en-us/library/cc236695.aspx
is not very clear.