Implement SMB3 Signing (AesCmac), 3.0.x encryption (AES/CCM) and 3.1.1 encryption (AES/GCM) (#438)
* Implement SMB3 Signing (AesCmac), 3.0.x encryption (AES/CCM) and 3.1.1 encryption (AES/GCM)

Fix Cancel Request didn't send sessionId. Added NegotitateConext classes for 3.1.1 and applied them on negotiate request and response. Added transformHeader for SMB3 encryption and its packet factory for decrypt. Added PacketEncryptor for packet encryption. Added AEADBlockCipher for AES/CCM/NoPadding and AES/GCM/NoPadding, and this may integrate with the old cipher interface in the future. Handle the session and treeConnect flags on response for requiring encryption, so that, the library is able using plainText to access other share but only required encrypt share will be encrypted. Added AesCmac support in PacketSignatory for SMB3 signing. CryptographicKeysGenerator is added for key generate but not yet passed the security provider to it and always using BC classes directly. DigestUtil is added for 3.1.1 preAuthHashValue and the session authenticate process had always changed in order to calculate the preAuthHashValue for 3.1.1.

* Add test case for the SMB3 encryption classes. Dialect 3.1.1 test case is not yet added.

Add test case for KeyGenerator, PacketEncryptor, Smb3EncryptedPacketFactory and the transform header helper functions. Dialect 3.1.1 test case is not yet added.